+++ crates/biome_cli/src/lib.rs
@@ -194,14 +194,7 @@ impl<'app> CliSession<'app> {
                 cli_options,
                 write,
                 sub_command,
-            } => commands::migrate::migrate(
-                self,
-                cli_options,
-                write,
-                sub_command
-                    .map(|sub_command| sub_command.is_prettier())
-                    .unwrap_or_default(),
-            ),
+            } => commands::migrate::migrate(self, cli_options, write, sub_command),
             BiomeCommand::Search {
                 cli_options,
                 files_configuration,