+++ crates/biome_js_analyze/src/lint/style/no_restricted_globals.rs
@@ -62,7 +62,7 @@ const RESTRICTED_GLOBALS: [&str; 2] = ["event", "error"];
 pub struct RestrictedGlobalsOptions {
     /// A list of names that should trigger the rule
     #[serde(skip_serializing_if = "Vec::is_empty")]
-    denied_globals: Vec<String>,
+    pub denied_globals: Vec<String>,
 }
 
 impl Rule for NoRestrictedGlobals {